+++ src/main/java/com/verdantartifice/thaumicwonders/ThaumicWonders.java
@@ -59,9 +59,6 @@ public class ThaumicWonders {
     //  Thauma Llama
     //  Check out Gadomancy to see if any additions can be made to TW (specifically the Infusion Claw for infusion automation)
     //  Heximite explosion converts blocks into tainted variants (add CrT/GS integration for additional conversions)
-    //  Golem Seal "Research Assistant"
-    //      refills research table paper
-    //      searches for items needed for research
     //  Research Brews + Crucible addon - create brews that randomly grant research after being consumed.
     //  Teleport Home focus effect
     //  Portal Generator instability events and custom summons