+++ src/main/java/com/verdantartifice/thaumicwonders/common/blocks/devices/BlockDimensionalRipper.java
@@ -49,10 +49,10 @@ public class BlockDimensionalRipper extends BlockDeviceTW<TileDimensionalRipper>
     @Override
     public int getComparatorInputOverride(IBlockState blockState, World worldIn, BlockPos pos) {
         TileEntity tile = worldIn.getTileEntity(pos);
-        if(tile instanceof TileDimensionalRipper) {
+        if (tile instanceof TileDimensionalRipper) {
             int essentia = ((TileDimensionalRipper) tile).getAmount();
             int capacity = ((TileDimensionalRipper) tile).getCapacity();
-            if(essentia > 0) {
+            if (essentia > 0) {
                 return Math.max(1, (int) (15f * (float) essentia / (float) capacity));
             }
         }