+++ src/main/java/com/verdantartifice/thaumicwonders/core/ModMixins.java
@@ -1,0 +1,13 @@
+package com.verdantartifice.thaumicwonders.core;
+
+import zone.rong.mixinbooter.ILateMixinLoader;
+
+import java.util.Collections;
+import java.util.List;
+
+public class ModMixins implements ILateMixinLoader {
+    @Override
+    public List<String> getMixinConfigs() {
+        return Collections.singletonList("mixins.thaumicwonders.json");
+    }
+}